+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changes from version 0.8.2 to master 
 
 - Added `--docker.imagePullPolicy` option
+- Allow multiple `--starter.join` arguments. 
+- Fixed high CPU load (#75)
 
 # Changes from version 0.8.1 to master